Move RouteNotFoundException into Exceptions namespace
